backend/requirements.txt
fastapi
pydantic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography] 
